+++ python/acl_anthology/collections/volume.py
@@ -22,6 +22,7 @@ from typing import Any, Iterator, Optional, cast, TYPE_CHECKING
 import sys
 
 from .. import constants
+from ..config import config
 from ..containers import SlottedDict
 from ..files import PDFReference
 from ..people import NameSpecification
@@ -111,6 +112,11 @@ class Volume(SlottedDict[Paper]):
         return (self.parent.id, self.id, None)
 
     @property
+    def has_abstracts(self) -> bool:
+        """True if at least one paper in this volume has an abstract."""
+        return any(paper.abstract is not None for paper in self.data.values())
+
+    @property
     def has_frontmatter(self) -> bool:
         """True if this volume has frontmatter."""
         return "0" in self.data
@@ -125,6 +131,11 @@ class Volume(SlottedDict[Paper]):
     def root(self) -> Anthology:
         """The Anthology instance to which this object belongs."""
         return self.parent.parent.parent
+
+    @property
+    def web_url(self) -> str:
+        """The URL of this volume's landing page on the ACL Anthology website."""
+        return cast(str, config["volume_page_template"]).format(self.full_id)
 
     def get_events(self) -> list[Event]:
         """