+++ python/acl_anthology/utils/text.py
@@ -13,7 +13,23 @@
 # limitations under the License.
 
 
-from typing import cast
+from typing import cast, Optional
+
+
+_MONTH_TO_NUM: dict[str, int] = {
+    "january": 1,
+    "february": 2,
+    "march": 3,
+    "april": 4,
+    "may": 5,
+    "june": 6,
+    "july": 7,
+    "august": 8,
+    "september": 9,
+    "october": 10,
+    "november": 11,
+    "december": 12,
+}
 
 
 def interpret_pages(text: str) -> tuple[str, str]:
@@ -29,6 +45,21 @@ def interpret_pages(text: str) -> tuple[str, str]:
         if text.count(s) == 1:
             return cast(tuple[str, str], tuple(text.split(s)))
     return (text, text)
+
+
+def month_str2num(text: str) -> Optional[int]:
+    """Convert a month string to a number, e.g. February -> 2
+
+    Arguments:
+        text: A text string representing a month value.
+
+    Returns:
+        None if the string doesn't correspond to a month; the numeric month value otherwise.
+
+    Note:
+        We're not using Python's datetime here since its behaviour depends on the system locale.
+    """
+    return _MONTH_TO_NUM.get(text.lower(), None)
 
 
 def remove_extra_whitespace(text: str) -> str: